Clean more clang-tidy checks

+++ src/graphic/gl/utils.cc
@@ -18,6 +18,7 @@
 #include "graphic/gl/utils.h"
 
 #include <cassert>
+#include <cstddef>
 #include <memory>
 
 #include "base/wexception.h"
@@ -266,7 +267,7 @@ void vertex_attrib_pointer(int vertex_index, int num_items, int stride, size_t o
 
 void swap_rows(const int width, const int height, const int pitch, const int bpp, uint8_t* pixels) {
 	uint8_t* begin_row = pixels;
-	uint8_t* end_row = pixels + pitch * (height - 1);
+	uint8_t* end_row = pixels + static_cast<ptrdiff_t>(pitch * (height - 1));
 	while (begin_row < end_row) {
 		for (int x = 0; x < width * bpp; ++x) {
 			std::swap(begin_row[x], end_row[x]);
