+++ src/wui/attack_window.cc
@@ -18,6 +18,7 @@
 
 #include "wui/attack_window.h"
 
+#include <cstddef>
 #include <memory>
 
 #include <SDL_mouse.h>
@@ -459,7 +460,7 @@ bool AttackWindow::handle_mousewheel(int32_t x, int32_t y, uint16_t modstate) {
 Widelands::Extent AttackWindow::ListOfSoldiers::size() const {
 	const size_t nr_soldiers = count_soldiers();
 	uint32_t rows = nr_soldiers / current_size_;
-	if (nr_soldiers == 0 || rows * current_size_ < nr_soldiers) {
+	if (nr_soldiers == 0 || static_cast<unsigned long>(rows) * current_size_ < nr_soldiers) {
 		++rows;
 	}
 	if (restricted_row_number_) {
